Clean and Resample Species Data

This function cleans up the catch data for a specific species and then performs resampling tests,

clean_and_resample(
  spp_info,
  catch,
  seq_from,
  seq_to,
  seq_by,
  tot_dataframes,
  replicate_num,
  grid_yrs,
  dir_out,
  test = FALSE,
  n_knots = 300,
  model_type = "wrapper_sdmtmb"
)

Arguments

spp_info

A data frame row containing information about the species.

catch

A data frame containing the catch data.

seq_from

A numeric value specifying the start of the sequence for data frames.

seq_to

A numeric value specifying the end of the sequence for data frames.

seq_by

A numeric value specifying the step size of the sequence for data frames.

tot_dataframes

An integer specifying the total number of data frames to generate.

replicate_num

An integer specifying the number of replicates.

grid_yrs

A data frame or list containing grid years information.

dir_out

A character string specifying the directory for output files.

test

Logical. TRUE/FALSE. If TRUE, will only run first two resampling tests.

n_knots

Numeric. Default = 500.

model_type

String. Default = "wrapper_sdmtmb", but can be any preset wrapper_*() function or a premade home built function.

Value

A list of data frames containing the cleaned and resampled catch data.

Details

This function performs the following steps:

Cleans up the catch data for the specified species using `cleanup_by_species`.

Performs resampling tests on the cleaned data using `resample_tests`.

Examples