Cleanup species function

Filter catch by species, create a vector of tows and give tows a random assignment to be resampled,

cleanup_by_species(
  catch,
  spp_info,
  seq_from = 0.1,
  seq_to = 1,
  seq_by = 0.1,
  tot_dataframes = 91,
  replicate_num = 10
)

Arguments

catch

full catch df

spp_info

A data frame containing information about the test species.

seq_from

A numeric value specifying the start of the sequence for data frames.

seq_to

A numeric value specifying the end of the sequence for data frames.

seq_by

A numeric value specifying the step size of the sequence for data frames.

tot_dataframes

the number of data frames you want to output. effort x replicates - (replicates - 1). 5x3-2

replicate_num

An integer specifying the number of replicates.

x

catch_split data frame

Join Data Frames

Function to join a list of data frames to a main data frame using a shared column.

list_of_dfs

The list of data frames that you would like to join to the main data frame

main_df

The main data frame that you want to join the list of data frames

shared_column

The column that all the data frames share which will be used to join the data frames together.

Value

List of resampled catch data frames

Tow Function

Create a vector of tows for including or excluding.