AFSC biomass and population estimates data

The final, validated survey data are publicly accessible soon after surveys are completed on the Fisheries One Stop Shop (FOSS) platform. This data includes catch, haul, and environmental data collected at each station. On the FOSS data platform, users can interactively select, view, and download data. Descriptive documentation and user-examples are available on the metadata page.

data('noaa_afsc_biomass')

Format

A data frame with 738 observations on the following 8 variables.

species_code

Taxon scientific name. The species code of the organism associated with the common_name and scientific_name columns.

year

Survey year. Year the observation (survey) was collected.

biomass_mt

Estimated biomass. The estimated total biomass.

biomass_var

Estimated biomass variance. The estimated variance associated with the total biomass.

population_count

Estimated population. The estimated population caught in the survey for a species, group, or total for a given survey.

population_var

Estimated population variance. The estimated population variance caught in the survey for a species, group, or total for a given survey.

survey_definition_id

Survey ID. The survey definition ID key code is an integer that uniquely identifies a survey region/survey design. The column survey_definition_id is associated with the srvy and survey columns.

area_id

Area ID. Area ID key code for each statistical area used to produce production estimates (e.g., biomass, population, age comps, length comps). Each area ID is unique within each survey.

#'

Source

https://github.com/afsc-gap-products/gap_products

Details

The Resource Assessment and Conservation Engineering (RACE) Division Groundfish Assessment Program (GAP) of the Alaska Fisheries Science Center (AFSC) conducts fisheries-independent bottom trawl surveys to assess the populations of demersal fish and crab stocks of Alaska.

Author

Alaska Fisheries Science Center, compiled by Emily Markowitz (Emily.Markowitz AT noaa.gov)

Examples

data(noaa_afsc_biomass)