Plot Results and Save Figures

This function compiles results from species distribution models (SDMs) output of the clean_and_resample function for which it generates plots, and saves the figures.

plot_results(srvy, dir_out, dir_final = NULL)

Arguments

srvy

A character string specifying the survey identifier.

dir_out

A character string specifying the directory for output files.

dir_final

A character string specifying the directory for new result files.

Value

A list of plots and tables generated during the function execution.

Details

This function performs the following steps:

Creates a directory for saving images.

Searches for relevant files in the output directory.

Compiles data from the found files.

Saves the compiled data into CSV files.

Generates various plots such as boxplots and time series of biomass estimates.

TO DO: CREATE BIO PLOTS

Saves the generated plots as PNG files.

Saves the list of plots in an RData file.

Examples

dir_out <- here::here("vignettes", "output")
srvy <- "CA"
plot_results(srvy = srvy, dir_out = dir_out) 
#> Error in fit_df %>% dplyr::bind_rows(utils::read.csv(paste0(aaa[i], "/fit_df.csv"))): could not find function "%>%"